crowdfund/requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-countries==7.2.1
django-environ==0.8.1
django-phonenumber-field==6.0.0
django-querycount==0.7.0
django-rest-knox==4.1.0
django-taggit==2.0.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
environ==1.0
idna==3.3
oauthlib==3.1.1
phonenumbers==8.12.37
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
